requirements.txt
Django==1.11.3
django-sendfile==0.3.11
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
numpy==1.13.1
openpyxl==2.4.8
pandas==0.20.3
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
whitenoise==3.3.0
xlrd==1.0.0
xlwt==1.2.0
